chat-service/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
uvloop==0.19.0
httptools==0.6.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async
aiofiles==23.2.1
httpx==0.25.2
redis[hiredis]==5.0.1

# LangChain
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.10.0

# Vector storage
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Document processing
pdfplumber==0.9.0
python-docx==1.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
pytesseract==0.3.10
chardet==5.2.0
openpyxl==3.1.2
python-pptx==0.6.23
docxtpl==0.16.7

# NLP
razdel==0.5.0

# Utils
tqdm==4.66.1
python-json-logger==2.0.7